Read Description for Full Changes
Vehicle health is no longer being changed until all parts have been chopped, but measures have been put in place to ensure that players cannot take the vehicle back and repair it and continue chopping the parts, its not perfect but I'll be finding some method later where the cars don't explode.

Now tracking the vehicle ID from the spawned vehicle and using that ID to remove different parts and use the QBCore delete vehicle function.

Vehicles should now properly delete.

A few vehicle part amounts have been updated.

Reverted back to QB-Drawtext for the time being, the link to the repo will be in the readme, when its officially baked into the QBCore framework I'll move it to that.  Still looking into some popular drawtexts so that I can have a config setting for it. For the time being, if you don't know how to change the exports I recommend utilizing QB-Drawtext.